refactor: move selection component into the body component
