+++ projects/ngx-datatable/src/lib/components/body/scroller.component.ts
@@ -49,7 +49,7 @@ export class ScrollerComponent implements OnInit, OnDestroy {
     // manual bind so we don't always listen
     if (this.scrollbarV || this.scrollbarH) {
       const renderer = this.renderer;
-      this.parentElement = renderer.parentNode(renderer.parentNode(this.element));
+      this.parentElement = renderer.parentNode(this.element);
       this._scrollEventListener = this.onScrolled.bind(this);
       this.parentElement.addEventListener('scroll', this._scrollEventListener);
     }
